Allow ECC Verify-only operations with NO_ASN and SP_MATH
